opensupervision/requirements.txt
# OpenSupervision - Complete Open-Source Computer Vision Platform
# A sovereign alternative to Roboflow ecosystem

# Core Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ultralytics>=8.0.0  # YOLO models and training

# Inference and Model Serving
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0  # GPU acceleration
tensorrt>=8.6.0  # For TensorRT support (optional)

# Data Management
pyyaml>=6.0
jsonlines>=4.0
pandas>=2.0.0

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Visualization and UI
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0  # Training UI

# Experiment Tracking
mlflow>=2.4.0
wandb>=0.15.0  # Optional alternative to MLflow

# Dataset Management
dvc>=2.58.0  # Data Version Control
fiftyone>=0.20.0  # Visual dataset management (optional)
pymongo>=4.4.0  # For FiftyOne backend

# Annotation Tools Integration
requests>=2.31.0  # For CVAT API
beautifulsoup4>=4.12.0  # For web scraping (optional)

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
ffmpeg-python>=0.2.0

# Parallel Processing

joblib>=1.3.0

# Progress Tracking
tqdm>=4.65.0
rich>=13.4.0

# Configuration Management
python-dotenv>=1.0.0
click>=8.1.0

# Testing and Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Deployment and Containerization
docker>=6.1.0  # For containerization
gunicorn>=21.0.0  # Production server

# Advanced Tracking (Optional)
filterpy>=1.4.5  # Kalman filters
scikit-learn>=1.3.0  # Machine learning utilities
scikit-image>=0.21.0  # Advanced image processing

# Model Conversion and Optimization
openvino>=2023.0.0  # Intel optimizations (optional)
coremltools>=6.3.0  # Core ML export (optional)
tensorflow>=2.13.0  # Alternative DL framework (optional)

# Annotation and Labeling
labelme>=4.6.0  # Image annotation
cvat-sdk>=1.5.0  # CVAT API client (if not using requests directly)

# Storage Backends
boto3>=1.28.0  # AWS S3 (optional)
minio>=7.2.0  # MinIO client (optional)

# Monitoring and Logging
prometheus-client>=0.17.0  # Metrics collection
structlog>=23.1.0  # Structured logging

# Performance Optimization
numba>=0.57.0  # JIT compilation for performance
cython>=3.0.0  # C extensions

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Jupyter Support (Optional)
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Mobile and Edge Deployment (Optional)



# Additional Utilities
psutil>=5.9.0  # System monitoring
pathlib>=1.0.1  # Path handling
typing-extensions>=4.7.0  # Type hints

# Security
cryptography>=41.0.0
passlib>=1.7.0

# Note: Some packages may require system-level dependencies:
# - ffmpeg: For video processing
# - CUDA drivers: For GPU acceleration
# - OpenGL: For some visualization features
# - Node.js: For some frontend components (optional)

# Installation Notes:
# 1. For GPU support: Install CUDA-compatible versions of torch and onnxruntime-gpu
# 2. For TensorRT: Install TensorRT from NVIDIA (requires separate download)
# 3. For CVAT integration: Ensure CVAT server is running separately
# 4. For FiftyOne: Ensure MongoDB is running for backend
# 5. For MLflow: Can use SQLite for local tracking or set up PostgreSQL/MySQL

# Optional: Install with specific extras
# pip install opensupervision[gpu]  # GPU acceleration
# pip install opensupervision[full]  # All optional dependencies
# pip install opensupervision[dev]   # Development dependencies